IOPS-1870 Allow Van as surname when only surname
It was detected that Van was always detected as prefix of surname.
However, it did not make sense to have a prefix without a surname.
This PR fixes that, it checks if it is the only surname and if it
is, it get the only surname as surname without paying attention
to the prefixes.