Detect last name prefixes in combined lastname parts (fixes #30)
This enables the detection of lastnames combined via a dash with
the prefix to a following lastname, e.g. in "Etje Heijdanus-De Boer".
At this point the entire part is treated as Lastname as there is
no support for re-combining last name parts or prefixes.
This means the 'De' prefix in 'De Boer' above cannot be accessed
individually.